+++ packages/express/src/declarations.ts
@@ -2,15 +2,15 @@ import http from 'http';
 import express, { Express } from 'express';
 import {
   Application as FeathersApplication, Params as FeathersParams,
-  HookContext, ServiceMethods, ServiceInterface
+  HookContext, ServiceMethods, ServiceInterface, RouteLookup
 } from '@feathersjs/feathers';
 
-interface ExpressUseHandler<T, ServiceTypes> {
-  <L extends keyof ServiceTypes & string> (
+interface ExpressUseHandler<T, Services> {
+  <L extends keyof Services & string> (
     path: L,
     ...middlewareOrService: (
       Express|express.RequestHandler|
-      (keyof any extends keyof ServiceTypes ? ServiceInterface<any> : ServiceTypes[L])
+      (keyof any extends keyof Services ? ServiceInterface : Services[L])
     )[]
   ): T;
   (path: string|RegExp, ...expressHandlers: express.RequestHandler[]): T;
@@ -18,43 +18,46 @@ interface ExpressUseHandler<T, ServiceTypes> {
   (handler: Express|express.ErrorRequestHandler): T;
 }
 
-export interface ExpressOverrides<ServiceTypes> {
+export interface ExpressOverrides<Services> {
   listen(port: number, hostname: string, backlog: number, callback?: () => void): Promise<http.Server>;
   listen(port: number, hostname: string, callback?: () => void): Promise<http.Server>;
   listen(port: number|string|any, callback?: () => void): Promise<http.Server>;
   listen(callback?: () => void): Promise<http.Server>;
-  use: ExpressUseHandler<this, ServiceTypes>;
+  use: ExpressUseHandler<this, Services>;
+  server: http.Server;
 }
 
-export type Application<ServiceTypes = any, AppSettings = any> =
-  Omit<Express, 'listen'|'use'> &
-  FeathersApplication<ServiceTypes, AppSettings> &
-  ExpressOverrides<ServiceTypes>;
+export type Application<Services = any, Settings = any> =
+  Omit<Express, 'listen'|'use'|'get'|'set'> &
+  FeathersApplication<Services, Settings> &
+  ExpressOverrides<Services>;
 
 declare module '@feathersjs/feathers/lib/declarations' {
-  export interface ServiceOptions {
-    middleware?: {
-      before: express.RequestHandler[],
-      after: express.RequestHandler[]
-    }
+  interface ServiceOptions {
+    express?: {
+      before?: express.RequestHandler[];
+      after?: express.RequestHandler[];
+      composed?: express.RequestHandler;
+    };
   }
 }
 
 declare module 'express-serve-static-core' {
   interface Request {
-      feathers?: Partial<FeathersParams>;
+    feathers?: Partial<FeathersParams>;
+    lookup?: RouteLookup;
   }
 
   interface Response {
-      data?: any;
-      hook?: HookContext;
+    data?: any;
+    hook?: HookContext;
   }
 
   interface IRouterMatcher<T> {
-      // eslint-disable-next-line
-      <P extends Params = ParamsDictionary, ResBody = any, ReqBody = any>(
-          path: PathParams,
-          ...handlers: (RequestHandler<P, ResBody, ReqBody> | Partial<ServiceMethods<any, any>> | Application)[]
-      ): T;
+    // eslint-disable-next-line
+    <P extends Params = ParamsDictionary, ResBody = any, ReqBody = any>(
+        path: PathParams,
+        ...handlers: (RequestHandler<P, ResBody, ReqBody> | Partial<ServiceMethods> | Application)[]
+    ): T;
   }
 }