+++ packages/express/test/rest.test.ts
@@ -22,7 +22,7 @@ describe('@feathersjs/express/rest provider', () => {
       const app = feathers();
 
       try {
-        app.configure(rest());
+        app.configure(rest() as any);
         assert.ok(false, 'Should never get here');
       } catch (e: any) {
         assert.strictEqual(e.message, '@feathersjs/express/rest needs an Express compatible app.');
@@ -78,13 +78,12 @@ describe('@feathersjs/express/rest provider', () => {
   });
 
   describe('CRUD', () => {
-    let server: Server;
     let app: express.Application;
 
     before(async () => {
       app = expressify(feathers())
-        .configure(rest(express.formatter))
         .use(express.json())
+        .configure(rest(express.formatter))
         .use('codes', {
           async get (id: Id) {
             return { id };
@@ -97,10 +96,10 @@ describe('@feathersjs/express/rest provider', () => {
         .use('/', new Service())
         .use('todo', new Service());
 
-      server = await app.listen(4777, () => app.use('tasks', new Service()));
+      await app.listen(4777, () => app.use('tasks', new Service()));
     });
 
-    after(done => server.close(done));
+    after(() => app.teardown());
 
     restTests('Services', 'todo', 4777);
     restTests('Root Service', '/', 4777);
@@ -158,6 +157,7 @@ describe('@feathersjs/express/rest provider', () => {
           type: null,
           method: 'get',
           path: 'hook',
+          http: {},
           event: null,
           result: { description: 'You have to do dishes' },
           addedProperty: true
@@ -196,7 +196,7 @@ describe('@feathersjs/express/rest provider', () => {
 
         app.service('hook-status').hooks({
           after (hook: HookContext) {
-            hook.statusCode = 206;
+            hook.http.statusCode = 206;
           }
         });
 
@@ -245,7 +245,7 @@ describe('@feathersjs/express/rest provider', () => {
               event: null,
               method: 'get',
               path: 'hook-error',
-              original: data.hook.original
+              http: {}
             },
             error: { message: 'I blew up' }
           });
@@ -263,12 +263,12 @@ describe('@feathersjs/express/rest provider', () => {
       };
 
       const app = expressify(feathers())
-        .configure(rest(express.formatter))
         .use(function (req: Request, _res: Response, next: NextFunction) {
           assert.ok(req.feathers, 'Feathers object initialized');
           req.feathers.test = 'Happy';
           next();
         })
+        .configure(rest(express.formatter))
         .use('service', service);
       const server = await app.listen(4778);
 
@@ -300,8 +300,8 @@ describe('@feathersjs/express/rest provider', () => {
         req.headers['content-type'] = req.headers['content-type'] || 'application/json';
         next();
       })
-        .configure(rest(express.formatter))
         .use(express.json())
+        .configure(rest(express.formatter))
         .use('/todo', {
           async create (data: any) {
             return data;
@@ -325,8 +325,9 @@ describe('@feathersjs/express/rest provider', () => {
     it('allows middleware before and after a service', async () => {
       const app = expressify(feathers());
 
-      app.configure(rest())
+      app
         .use(express.json())
+        .configure(rest())
         .use('/todo', function (req, _res, next) {
           req.body.before = ['before first'];
           next();
@@ -360,8 +361,9 @@ describe('@feathersjs/express/rest provider', () => {
     it('allows middleware arrays before and after a service', async () => {
       const app = expressify(feathers());
 
-      app.configure(rest())
+      app
         .use(express.json())
+        .configure(rest())
         .use('/todo', [function (req: Request, _res: Response, next: NextFunction) {
           req.body.before = ['before first'];
           next();
@@ -404,8 +406,9 @@ describe('@feathersjs/express/rest provider', () => {
           res.data.push(req.body.text);
           res.status(200).json(res.data);
         }];
-      app.configure(rest())
+      app
         .use(express.json())
+        .configure(rest())
         .use('/array-middleware', middlewareArray);
 
       const server = await app.listen(4776);
@@ -575,8 +578,8 @@ describe('@feathersjs/express/rest provider', () => {
 
     before(async () => {
       app = expressify(feathers())
-        .configure(rest())
         .use(express.json())
+        .configure(rest())
         .use('/todo', new Service(), {
           methods: ['find', 'customMethod']
         })