+++ packages/feathers/src/service.ts
@@ -24,11 +24,13 @@ export const defaultEventMap = {
 export const protectedMethods = Object.keys(Object.prototype)
   .concat(Object.keys(EventEmitter.prototype))
   .concat([
+    'all',
     'before',
     'after',
     'error',
     'hooks',
     'setup',
+    'teardown',
     'publish'
   ]);