+++ packages/koa/src/declarations.ts
@@ -1,9 +1,10 @@
-import Koa from 'koa';
+import Koa, { Next } from 'koa';
 import { Server } from 'http';
-import { Application as FeathersApplication } from '@feathersjs/feathers';
+import { Application as FeathersApplication, HookContext, Params, RouteLookup } from '@feathersjs/feathers';
 import '@feathersjs/authentication';
 
 export type ApplicationAddons = {
+  server: Server;
   listen (port?: number, ...args: any[]): Promise<Server>;
 }
 
@@ -13,3 +14,23 @@ export type Application<T = any, C = any> =
 export type FeathersKoaContext<A = Application> = Koa.Context & {
   app: A;
 };
+
+export type Middleware<A = Application> = (context: FeathersKoaContext<A>, next: Next) => any;
+
+declare module '@feathersjs/feathers/lib/declarations' {
+  interface ServiceOptions {
+    koa?: {
+      before?: Middleware[];
+      after?: Middleware[];
+      composed?: Middleware;
+    };
+  }
+}
+
+declare module 'koa' {
+  interface ExtendableContext {
+    feathers?: Partial<Params>;
+    lookup?: RouteLookup;
+    hook?: HookContext;
+  }
+}