+++ packages/koa/src/rest.ts
@@ -1,51 +1,93 @@
-import { Next } from 'koa';
+import compose from 'koa-compose';
 import { http } from '@feathersjs/transport-commons';
 import { createDebug } from '@feathersjs/commons';
 import { getServiceOptions, defaultServiceMethods, createContext } from '@feathersjs/feathers';
 import { MethodNotAllowed } from '@feathersjs/errors';
-import { FeathersKoaContext } from './declarations';
 
-const debug = createDebug('@feathersjs/koa:rest');
+import { Application, Middleware } from './declarations';
+import { AuthenticationSettings, parseAuthentication } from './authentication';
 
-export function rest () {
-  return async (ctx: FeathersKoaContext, next: Next) => {
-    const { app, request } = ctx;
-    const { query: koaQuery, path, body: data, method: httpMethod } = request;
-    const query = { ...koaQuery };
-    const methodOverride = request.headers[http.METHOD_HEADER] ?
-      request.headers[http.METHOD_HEADER] as string : null;
-    const lookup = app.lookup(path);
+const debug = createDebug('@feathersjs/koa/rest');
 
-    if (lookup !== null) {
-      const { service, params: { __id: id = null, ...route } = {} } = lookup;
-      const method = http.getServiceMethod(httpMethod, id, methodOverride);
-      const { methods } = getServiceOptions(service);
-      
-      debug(`Found service for path ${path}, attempting to run '${method}' service method`);
+const serviceMiddleware = (): Middleware => {
+  return async (ctx, next) => {
+    const { query, headers, path, body: data, method: httpMethod } = ctx.request;
+    const methodOverride = ctx.request.headers[http.METHOD_HEADER] as string | undefined;
 
-      if (!methods.includes(method) || defaultServiceMethods.includes(methodOverride)) {
-        ctx.response.status = http.statusCodes.methodNotAllowed;
+    const { service, params: { __id: id = null, ...route } = {} } = ctx.lookup!;
+    const method = http.getServiceMethod(httpMethod, id, methodOverride);
+    const { methods } = getServiceOptions(service);
 
-        throw new MethodNotAllowed(`Method \`${method}\` is not supported by this endpoint.`);
-      }
+    debug(`Found service for path ${path}, attempting to run '${method}' service method`);
 
-      const createArguments = (http.argumentsFor as any)[method] || http.argumentsFor.default;
-      const params = {
-        ...ctx.feathers,
-        query,
-        route
-      };
-      const args = createArguments({ id, data, params });
-      const hookContext = createContext(service, method);
-
-      ctx.hook = hookContext as any;
-
-      const result = await (service as any)[method](...args, hookContext);
-
-      ctx.response.status = http.getStatusCode(result, {});
-      ctx.body = http.getData(result);
+    if (!methods.includes(method) || defaultServiceMethods.includes(methodOverride)) {
+      const error = new MethodNotAllowed(`Method \`${method}\` is not supported by this endpoint.`);
+      ctx.response.status = error.code;
+      throw error;
     }
+
+    const createArguments = http.argumentsFor[method as 'get'] || http.argumentsFor.default;
+    const params = { query, headers, route, ...ctx.feathers };
+    const args = createArguments({ id, data, params });
+    const contextBase = createContext(service, method, { http: {} });
+    ctx.hook = contextBase;
+
+    const context = await (service as any)[method](...args, contextBase);
+    ctx.hook = context;
+
+    const result = http.getData(context);
+    const statusCode = http.getStatusCode(context, result);
+
+    ctx.body = result;
+    ctx.status = statusCode;
 
     return next();
   };
-}
+};
+
+const servicesMiddleware = (): Middleware => {
+  return async (ctx, next) => {
+    const app = ctx.app;
+    const lookup = app.lookup(ctx.request.path);
+
+    if (!lookup) {
+      return next();
+    }
+
+    ctx.lookup = lookup;
+
+    const options = getServiceOptions(lookup.service);
+    const middleware = options.koa!.composed!;
+
+    return middleware(ctx, next);
+  };
+};
+
+export const formatter: Middleware = (_ctx, _next) => {};
+
+export type RestOptions = {
+  formatter?: Middleware;
+  authentication?: AuthenticationSettings;
+};
+
+export const rest = (options?: RestOptions | Middleware) => {
+  options = typeof options === 'function' ? { formatter: options } : options || {};
+
+  const formatterMiddleware = options.formatter || formatter;
+  const authenticationOptions = options.authentication;
+
+  return (app: Application) => {
+    app.use(parseAuthentication(authenticationOptions));
+    app.use(servicesMiddleware());
+
+    app.mixins.push((_service, _path, options) => {
+      const { koa: { before = [], after = [] } = {} } = options;
+
+      const middlewares = [].concat(before, serviceMiddleware(), after, formatterMiddleware);
+      const middleware = compose(middlewares);
+
+      options.koa ||= {};
+      options.koa.composed = middleware;
+    });
+  };
+};