+++ packages/koa/test/index.test.ts
@@ -1,20 +1,28 @@
 import { strict as assert } from 'assert';
 import Koa  from 'koa';
 import axios from 'axios';
-import { Server } from 'http';
 import { feathers, Id } from '@feathersjs/feathers';
 import { Service, restTests } from '@feathersjs/tests';
 import { koa, rest, Application, bodyParser, errorHandler } from '../src';
 
 describe('@feathersjs/koa', () => {
   let app: Application;
-  let server: Server;
 
   before(async () => {
     app = koa(feathers());
     app.use(errorHandler());
     app.use(bodyParser());
-    app.use(rest());
+    app.use(async (ctx, next) => {
+      if (ctx.request.path === '/middleware') {
+        ctx.body = {
+          feathers: ctx.feathers,
+          message: 'Hello from middleware'
+        };
+      } else {
+        await next();
+      }
+    });
+    app.configure(rest());
     app.use('/', new Service());
     app.use('todo', new Service(), {
       methods: [
@@ -22,19 +30,11 @@ describe('@feathersjs/koa', () => {
         'patch', 'remove', 'customMethod'
       ]
     });
-    app.use(ctx => {
-      if (ctx.request.path === '/middleware') {
-        ctx.body = {
-          feathers: ctx.feathers,
-          message: 'Hello from middleware'
-        };
-      }
-    });
 
-    server = await app.listen(8465);
+    await app.listen(8465);
   });
 
-  after(() => server.close());
+  after(() => app.teardown());
 
   it('throws an error when initialized with invalid application', () => {
     try {
@@ -51,7 +51,7 @@ describe('@feathersjs/koa', () => {
 
   it('Koa wrapped and context.app are the same', async () => {
     const app = koa(feathers());
-    
+
     app.use('/test', {
       async get (id: Id) {
         return { id };
@@ -97,7 +97,7 @@ describe('@feathersjs/koa', () => {
         'X-Service-Method': 'customMethod'
       }
     });
-    
+
     assert.deepStrictEqual(data, {
       data: { message: 'Custom hello' },
       method: 'customMethod',
@@ -136,6 +136,19 @@ describe('@feathersjs/koa', () => {
 
       return true;
     });
+  });
+
+  it('.teardown closes http server', async () => {
+    const app = koa(feathers());
+    let called = false;
+
+    const server = await app.listen(8787);
+    server.on('close', () => {
+      called = true;
+    })
+
+    await app.teardown();
+    assert.ok(called);
   });
 
   restTests('Services', 'todo', 8465);