+++ packages/schema/src/query.ts
@@ -1,7 +1,7 @@
 import { JSONSchema } from 'json-schema-to-ts';
 
 export const queryProperty = <T extends JSONSchema> (definition: T) => ({
-  oneOf: [
+  anyOf: [
     definition,
     {
       type: 'object',
@@ -23,12 +23,4 @@ export const queryProperty = <T extends JSONSchema> (definition: T) => ({
       }
     }
   ]
-} as const);
-
-export const queryArray = <T extends readonly string[]> (fields: T) => ({
-  type: 'array',
-  items: {
-    type: 'string',
-    enum: fields
-  }
 } as const);