+++ packages/schema/test/fixture.ts
@@ -2,11 +2,11 @@ import {
   feathers, HookContext, Application as FeathersApplication
 } from '@feathersjs/feathers';
 import { memory, Service } from '@feathersjs/memory';
+import { GeneralError } from '@feathersjs/errors';
 
 import {
   schema, resolve, Infer, resolveResult,
-  queryProperty, queryArray, resolveQuery,
-  validateQuery, validateData, resolveData
+  queryProperty, resolveQuery, resolveData
 } from '../src';
 
 export const userSchema = schema({
@@ -20,17 +20,23 @@ export const userSchema = schema({
   }
 } as const);
 
-export const userResultSchema = userSchema.extend({
+export const userResultSchema = schema({
   $id: 'UserResult',
+  type: 'object',
+  additionalProperties: false,
+  required: ['id', ...userSchema.definition.required],
   properties: {
+    ...userSchema.definition.properties,
     id: { type: 'number' }
   }
-});
+} as const);
 
 export type User = Infer<typeof userSchema>;
 export type UserResult = Infer<typeof userResultSchema>;
 
 export const userDataResolver = resolve<User, HookContext<Application>>({
+  schema: userSchema,
+  validate: 'before',
   properties: {
     password: async () => {
       return 'hashed';
@@ -39,6 +45,7 @@ export const userDataResolver = resolve<User, HookContext<Application>>({
 });
 
 export const userResultResolver = resolve<UserResult, HookContext<Application>>({
+  schema: userResultSchema,
   properties: {
     password: async (value, _user, context) => {
       return context.params.provider ? undefined : value;
@@ -57,11 +64,35 @@ export const messageSchema = schema({
   }
 } as const);
 
-export const messageResultSchema = messageSchema.extend({
+export const messageResultSchema = schema({
   $id: 'MessageResult',
+  type: 'object',
+  additionalProperties: false,
+  required: ['id', 'user', ...messageSchema.definition.required],
   properties: {
+    ...messageSchema.definition.properties,
     id: { type: 'number' },
     user: { $ref: 'UserResult' }
+  }
+} as const);
+
+export type Message = Infer<typeof messageSchema>;
+export type MessageResult = Infer<typeof messageResultSchema> & {
+  user: User;
+};
+
+export const messageResultResolver = resolve<MessageResult, HookContext<Application>>({
+  schema: messageResultSchema,
+  properties: {
+    user: async (_value, message, context) => {
+      const { userId } = message;
+
+      if (context.params.error === true) {
+        throw new GeneralError('This is an error');
+      }
+
+      return context.app.service('users').get(userId, context.params);
+    }
   }
 });
 
@@ -70,7 +101,6 @@ export const messageQuerySchema = schema({
   type: 'object',
   additionalProperties: false,
   properties: {
-    $resolve: queryArray(messageResultSchema.propertyNames),
     $limit: {
       type: 'number',
       minimum: 0,
@@ -78,6 +108,10 @@ export const messageQuerySchema = schema({
     },
     $skip: {
       type: 'number'
+    },
+    $resolve: {
+      type: 'array',
+      items: { type: 'string' }
     },
     userId: queryProperty({
       type: 'number'
@@ -88,32 +122,15 @@ export const messageQuerySchema = schema({
 export type MessageQuery = Infer<typeof messageQuerySchema>;
 
 export const messageQueryResolver = resolve<MessageQuery, HookContext<Application>>({
+  schema: messageQuerySchema,
+  validate: 'before',
   properties: {
-    $resolve: async (value) => {
-      return value || messageResultSchema.propertyNames;
-    },
-
     userId: async (value, _query, context) => {
       if (context.params?.user) {
         return context.params.user.id;
       }
 
       return value;
-    }
-  }
-});
-
-export type Message = Infer<typeof messageSchema>;
-export type MessageResult = Infer<typeof messageResultSchema> & {
-  user: User;
-};
-
-export const messageResultResolver = resolve<MessageResult, HookContext<Application>>({
-  properties: {
-    user: async (_value, message, context) => {
-      const { userId } = message;
-
-      return context.app.service('users').get(userId, context.params);
     }
   }
 });
@@ -131,7 +148,6 @@ const app = feathers<ServiceTypes>()
   .use('messages', memory());
 
 app.service('messages').hooks([
-  validateQuery(messageQuerySchema),
   resolveQuery(messageQueryResolver),
   resolveResult(messageResultResolver)
 ]);
@@ -142,7 +158,6 @@ app.service('users').hooks([
 
 app.service('users').hooks({
   create: [
-    validateData(userSchema),
     resolveData(userDataResolver)
   ]
 });