+++ packages/transport-commons/test/routing/index.test.ts
@@ -69,4 +69,37 @@ describe('app.routes', () => {
       second: '::special'
     });
   });
+
+  it('can register routes with preset params', () => {
+    app.routes.insert('/my/service/:__id/preset', {
+      service: app.service('/my/service'),
+      params: { preset: true }
+    });
+
+    const result = app.lookup('/my/service/1234/preset');
+
+    assert.strictEqual(result.service, app.service('/my/service'));
+    assert.deepStrictEqual(result.params, {
+      preset: true,
+      __id: '1234'
+    });
+  });
+
+  it('can pass route params during a service registration', () => {
+    app.use('/other/service', {
+      async get (id: any) {
+        return id;
+      }
+    }, {
+      routeParams: { used: true }
+    });
+
+    const result = app.lookup('/other/service/1234');
+
+    assert.strictEqual(result.service, app.service('/other/service'));
+    assert.deepStrictEqual(result.params, {
+      used: true,
+      __id: '1234'
+    });
+  });
 });